requirements.txt
qcloudapi-sdk-python
peewee
PyMySQL
flask
futures